update payment signals to use the django 1.7 style - decorators in a separate signals.py file, imported in __init__
