+++ bluebottle/bb_orders/models.py
@@ -88,4 +88,3 @@ class BaseOrder(models.Model, FSMTransition):
         preview_serializer = 'bluebottle.bb_orders.serializers.OrderSerializer'
         manage_serializer = 'bluebottle.bb_orders.serializers.ManageOrderSerializer'
 
-from .signals import *
